+++ R/movecorr.R
@@ -7,7 +7,7 @@
 #' the least cost corridor (with least-cost paths superimposed), and returns a list containing a number of components (see 'Value' below). The corridor raster can be
 #' exported as GeoTiff (see 'Arguments' below). \cr
 #'
-#' What 'movecorr' does is to calculate (via the \code{\link{movecost}} function) the accumulated cost surface around location
+#' What 'movecorr()' does is to calculate (via the \code{\link{movecost}} function) the accumulated cost surface around location
 #' a, and then the accumulated cost surface around b. The two surfaces are eventually combined to produce the least-cost
 #' corridor between location a and b. On the produced corridor raster, the cost of a cell is the total cost to reach it
 #' from both locations. About least-cost corridors see for instance: \cr
@@ -17,8 +17,11 @@
 #' @param dtm Digital Terrain Model (RasterLayer class); if not provided, elevation data will be acquired online for the area enclosed by the 'studyplot' parameter (see \code{\link{movecost}}).
 #' @param a first location from which the least-cost corridor is calculated (SpatialPointsDataFrame class).
 #' @param b second location from which the least-cost corridor is calculated (SpatialPointsDataFrame class).
+#' @param lab.a string to be used to label point a on the outplut plot (A is the default)
+#' @param lab.b string to be used to label point a on the outplut plot (B is the default).
+#' @param cex.labs scaling factor for the size of the points' labels (0.8 by default)
 #' @param studyplot polygon (SpatialPolygonDataFrame class) representing the study area for which online elevation data are aquired (see \code{\link{movecost}}); NULL is default.
-#' @param funct cost function to be used:\cr \strong{t} (default) uses the on-path Tobler's hiking function;\cr
+#' @param funct cost function to be used (for details on each of the following, see \code{\link{movecost}}):\cr \strong{t} (default) uses the on-path Tobler's hiking function;\cr
 #' \strong{tofp} uses the off-path Tobler's hiking function;\cr
 #' \strong{mp} uses the Marquez-Perez et al.'s modified Tobler's function;\cr
 #' \strong{icmonp} uses the Irmischer-Clarke's modified Tobler's hiking function (male, on-path);\cr
@@ -32,24 +35,26 @@
 #' \strong{wcs} uses the wheeled-vehicle critical slope cost function;\cr
 #' \strong{p} uses the Pandolf et al.'s metabolic energy expenditure cost function;\cr
 #' \strong{vl} uses the Van Leusen's metabolic energy expenditure cost function;\cr
-#' \strong{ls} uses the Llobera-Sluckin's metabolic energy expenditure cost function (see Details).
-#' @param time time-unit expressed by the accumulated raster and by the isolines if Tobler's and Tobler-related cost functions are used;
+#' \strong{ls} uses the Llobera-Sluckin's metabolic energy expenditure cost function.
+#' @param time time-unit expressed by the accumulated raster if Tobler's and other time-related cost functions are used;
 #' 'h' for hour, 'm' for minutes.
 #' @param move number of directions in which cells are connected: 4 (rook's case), 8 (queen's case), 16 (knight and one-cell queen moves; default).
-#' @param sl.crit critical slope (in percent), typically in the range 8-16 (10 by default) (used by the wheeled-vehicle cost function; see Details).
-#' @param W walker's body weight (in Kg; 70 by default; used by the Pandolf's and Van Leusen's cost function; see Details).
-#' @param L carried load weight (in Kg; 0 by default; used by the Pandolf's and Van Leusen's cost function; see Details).
-#' @param N coefficient representing ease of movement (1 by default) (used by the Pandolf's and Van Leusen's cost function; see Details).
-#' @param V speed in m/s (1.2 by default) (used by the Pandolf's and Van Leusen's cost function; see Details).
-#' @param z zoom level for the elevation data downloaded from online sources (from 0 to 15; 9 by default) (see Details and \code{\link[elevatr]{get_elev_raster}}).
-#' @param export TRUE or FALSE (default) if the user wants or does not want the output to be exported; if TRUE, the least-cost corridor will be
-#' exported as a GeoTiff file.
+#' @param sl.crit critical slope (in percent), typically in the range 8-16 (10 by default) (used by the wheeled-vehicle cost function; see \code{\link{movecost}}).
+#' @param W walker's body weight (in Kg; 70 by default; used by the Pandolf's and Van Leusen's cost function; see \code{\link{movecost}}).
+#' @param L carried load weight (in Kg; 0 by default; used by the Pandolf's and Van Leusen's cost function; see \code{\link{movecost}}).
+#' @param N coefficient representing ease of movement (1 by default) (used by the Pandolf's and Van Leusen's cost function; see \code{\link{movecost}}).
+#' @param V speed in m/s (1.2 by default) (used by the Pandolf's and Van Leusen's cost function; see \code{\link{movecost}}).
+#' @param z zoom level for the elevation data downloaded from online sources (from 0 to 15; 9 by default) (see \code{\link{movecost}} and \code{\link[elevatr]{get_elev_raster}}).
+#' @param export TRUE or FALSE (default) if the user wants or does not want the output to be exported; if TRUE, the least-cost corridor, the dtm (if not provided by the user but acquired online),
+#' and the accumulated cost surface around a and b are exported as a GeoTiff file; the two LCPs (from a to b, and from b to a) as individual shapefiles.
 #'
 #' @return The function returns a list storing the following components \itemize{
 ##'  \item{dtm: }{Digital Terrain Model ('RasterLayer' class)}
 ##'  \item{lc.corridor: }{raster of the least-cost corridor ('RasterLayer' class)}
 ##'  \item{lcp_a_to_b: }{least-cost past from a to b ('SpatialLines' class)}
 ##'  \item{lcp_b_to_a: }{least-cost past from b to a ('SpatialLines' class)}
+##'  \item{accum_cost_surf_a: }{accumulated cost-surface around a ('RasterLayer' class)}
+##'  \item{accum_cost_surf_b: }{accumulated cost-surface around b ('RasterLayer' class)}
 ##' }
 ##'
 #' @keywords movecorr
@@ -69,13 +74,13 @@
 #' # relative energetic expenditure cost function, and store the results
 #' # in the 'res' object
 #'
-#' res <- movecorr(dtm=volc, a=destin.loc[1,], b=destin.loc[3,], funct="ree")
+#' result <- movecorr(dtm=volc, a=destin.loc[1,], b=destin.loc[3,], funct="ree")
 #'
 #'
 #' @seealso \code{\link{movecost}}
 #'
 #'
-movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=16, sl.crit=10, W=70, L=0, N=1, V=1.2, z=9, export=FALSE){
+movecorr <- function (dtm=NULL, a, b, lab.a="A", lab.b="B", cex.labs=0.8, studyplot=NULL, funct="t", time="h", move=16, sl.crit=10, W=70, L=0, N=1, V=1.2, z=9, export=FALSE){
 
   #calculate the accum cost surface and LCP around and from point a
   res.a <- movecost(dtm=dtm, origin=a, destin=b, studyplot=studyplot, funct=funct, time=time, move=move, sl.crit=sl.crit, W=W, L=L, N=N, V=V, z=z, graph.out=FALSE)
@@ -92,7 +97,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if (funct=="t") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the Tobler's on-path hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -100,7 +105,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if (funct=="tofp") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the Tobler's off-path hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -108,7 +113,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="mp") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the Marquez-Perez et al.'s modified Tobler hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -116,7 +121,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="icmonp") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the (male, on-path) Irmischer-Clarke's modified Tobler hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -124,7 +129,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="icmoffp") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the (male, off-path) Irmischer-Clarke's modified Tobler hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -132,13 +137,13 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="icfonp") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the (female, on-path) Irmischer-Clarke's modified Tobler hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
 
   if(funct=="icfoffp") {
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the (female, off-path) Irmischer-Clarke's modified Tobler hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -146,7 +151,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="ug") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the Uriarte Gonzalez's cost function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -154,7 +159,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="ree") {
 
     #set the labels to be used within the returned plot
-    main.title <- "Least-cost corridor between A and B"
+    main.title <- "Least-cost corridor"
     sub.title <- "Cost based on the slope-based relative energetic expenditure cost function"
     legend.cost <- "cost"
   }
@@ -162,7 +167,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="hrz") {
 
     #set the labels to be used within the returned plot
-    main.title <- "Least-cost corridor between A and B"
+    main.title <- "Least-cost corridor"
     sub.title <- "Cost based on the Herzog's metabolic cost function \n cost in J / (Kg*m)"
     legend.cost <- "metabolic cost J / (Kg*m)"
   }
@@ -170,7 +175,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="wcs") {
 
     #set the labels to be used within the returned plot
-    main.title <- "Least-cost corridor between A and B"
+    main.title <- "Least-cost corridor"
     sub.title <- paste0("Cost based on the wheeled-vehicle critical slope cost function \ncritical slope set to ", sl.crit, " percent")
     legend.cost <- "cost"
   }
@@ -178,7 +183,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="vl") {
 
     #set the labels to be used within the returned plot
-    main.title <- "Least-cost corridor between A and B"
+    main.title <- "Least-cost corridor"
     sub.title <- paste0("Cost based on the Van Leusen's metabolic energy expenditure cost function \nparameters: W: ", W, "; L: ", L, "; N: ", N, "; V: ", V)
     legend.cost <- "energy expenditure cost (Megawatts)"
   }
@@ -186,7 +191,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="p") {
 
     #set the labels to be used within the returned plot
-    main.title <- "Least-cost corridor between A and B"
+    main.title <- "Least-cost corridor"
     sub.title <- paste0("Cost based on the Pandolf et al.'s metabolic energy expenditure cost function \nparameters: W: ", W, "; L: ", L, "; N: ", N, "; V: ", V)
     legend.cost <- "energy expenditure cost (Megawatts)"
   }
@@ -194,7 +199,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="ls") {
 
     #set the labels to be used within the returned plot
-    main.title <- "Least-cost corridor between A and B"
+    main.title <- "Least-cost corridor"
     sub.title <- paste0("Cost based on the Llobera-Sluckin's metabolic energy expenditure cost function")
     legend.cost <- "energy expenditure cost (KJ/m)"
   }
@@ -202,7 +207,7 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
   if(funct=="alb") {
 
     #set the labels to be used within the returned plot
-    main.title <- paste0("Least-cost corridor (cost in ", time, ") between A and B")
+    main.title <- paste0("Least-cost corridor (cost in ", time, ")")
     sub.title <- "Walking-time based on the Alberti's modified Tobler hiking function"
     legend.cost <- paste0("walking-time (", time,")")
   }
@@ -238,26 +243,38 @@ movecorr <- function (dtm=NULL, a, b, studyplot=NULL, funct="t", time="h", move=
 
   #add label to point a
   raster::text(sp::coordinates(a),
-               labels="A",
+               labels=lab.a,
                col="red",
                pos = 4,
-               cex=0.8)
+               cex=cex.labs)
 
   #add label to point b
   raster::text(sp::coordinates(b),
-               labels="B",
+               labels=lab.b,
                pos = 4,
-               cex=0.8)
+               cex=cex.labs)
 
 
   #if export is TRUE, export the LC corridor as a raster file
+  #and the LCPs as shapefile
   if(export==TRUE){
     raster::writeRaster(res.corridor, paste0("LCcorridor_", funct), format="GTiff")
+    raster::writeRaster(res.a$accumulated.cost.raster, paste0("Accum_cost_surf_a_", funct), format="GTiff")
+    raster::writeRaster(res.b$accumulated.cost.raster, paste0("Accum_cost_surf_b_", funct), format="GTiff")
+    rgdal::writeOGR(res.a$LCPs, ".", paste0("lcp_a_to_b_", funct), driver="ESRI Shapefile")
+    rgdal::writeOGR(res.b$LCPs, ".", paste0("lcp_b_to_a_", funct), driver="ESRI Shapefile")
   }
 
-  results <- list("dtm"=dtm,
+  #if no DTM was provided (i.e., if 'studyplot' is not NULL), export the downloaded DTM as a raster file
+  if(export==TRUE & is.null(studyplot)==FALSE){
+    raster::writeRaster(res.a$dtm, "dtm", format="GTiff")
+  }
+
+  results <- list("dtm"=res.a$dtm,
                   "lc.corridor"=res.corridor,
                   "lcp_a_to_b"=res.a$LCPs,
-                  "lcp_b_to_a"=res.b$LCPs)
+                  "lcp_b_to_a"=res.b$LCPs,
+                  "accum_cost_surf_a"=res.a$accumulated.cost.raster,
+                  "accum_cost_surf_b"=res.b$accumulated.cost.raster)
 
   }